refactor(camel-plugin): Separate component with input and suggestions to the standalone component
